Don't remove energy from the PSS when it attempts to fill a full hatch
should resolve #13317 in GTNewHorizons/GT-New-Horizons-Modpack, where the issues was mistakenly identified as the PSS accepting power when it is full. In fact, it does stop accepting power when full, but it is constantly leaking power into hatches.

However it should be noted that unless the PSS has no hatches, it pays a tax each tick which will of course mean that it is almost never going to be filled up.